fix bugs

+++ app/models/model_thread.go
@@ -17,7 +17,7 @@ import (
 type Thread struct {
 
 	// Идентификатор ветки обсуждения.
-	Id float32 `json:"id,omitempty"`
+	Id int `json:"id,omitempty"`
 
 	// Заголовок ветки обсуждения.
 	Title string `json:"title"`
@@ -32,7 +32,7 @@ type Thread struct {
 	Message string `json:"message"`
 
 	// Кол-во голосов непосредственно за данное сообщение форума.
-	Votes float32 `json:"votes,omitempty"`
+	Votes int `json:"votes,omitempty"`
 
 	// Человекопонятный URL (https://ru.wikipedia.org/wiki/%D0%A1%D0%B5%D0%BC%D0%B0%D0%BD%D1%82%D0%B8%D1%87%D0%B5%D1%81%D0%BA%D0%B8%D0%B9_URL). В данной структуре slug опционален и не может быть числом.
 	Slug string `json:"slug,omitempty"`
